requirements.txt
flask~=3.1.0
flask-cors~=6.0.1
python-dotenv~=1.2.1
requests~=2.32.5
langchain~=1.0.3
langchain-google-genai~=3.0.0
langgraph~=1.0.2
langchain-core~=1.0.2
beautifulsoup4~=4.12.3
deepagents~=0.2.4
pydantic~=2.12.3
gunicorn~=21.2.0